Folder structure rework
