+++ src/main/java/com/kireiiiiiiii/shooting_stars/Game.java
@@ -31,8 +31,8 @@ import java.io.IOException;
 import java.util.Arrays;
 import java.util.ArrayList;
 
-import com.kireiiiiiiii.shooting_stars.ui.Interactable;
 import com.kireiiiiiiii.shooting_stars.common.AdvancedVariable;
+import com.kireiiiiiiii.shooting_stars.common.GPanel;
 import com.kireiiiiiiii.shooting_stars.common.PausableTimer;
 import com.kireiiiiiiii.shooting_stars.common.Settings;
 import com.kireiiiiiiii.shooting_stars.common.Vec2D;
@@ -45,25 +45,25 @@ import com.kireiiiiiiii.shooting_stars.constants.Keybinds;
 import com.kireiiiiiiii.shooting_stars.constants.Logs;
 import com.kireiiiiiiii.shooting_stars.constants.Textures;
 import com.kireiiiiiiii.shooting_stars.constants.WidgetTags;
+import com.kireiiiiiiii.shooting_stars.interfaces.Interactable;
+import com.kireiiiiiiii.shooting_stars.interfaces.Renderable;
 import com.kireiiiiiiii.shooting_stars.tools.ScreenUtil;
-import com.kireiiiiiiii.shooting_stars.ui.GPanel;
-import com.kireiiiiiiii.shooting_stars.ui.Renderable;
-import com.kireiiiiiiii.shooting_stars.ui.elements.Backround;
-import com.kireiiiiiiii.shooting_stars.ui.elements.game.GameOverScreen;
-import com.kireiiiiiiii.shooting_stars.ui.elements.game.HomeButton;
-import com.kireiiiiiiii.shooting_stars.ui.elements.game.PauseScreen;
-import com.kireiiiiiiii.shooting_stars.ui.elements.game.ScoreBoard;
-import com.kireiiiiiiii.shooting_stars.ui.elements.game.ScoreWidget;
-import com.kireiiiiiiii.shooting_stars.ui.elements.game.StarWidget;
-import com.kireiiiiiiii.shooting_stars.ui.elements.game.TimerWidget;
-import com.kireiiiiiiii.shooting_stars.ui.elements.game.TopscoreWidget;
-import com.kireiiiiiiii.shooting_stars.ui.elements.menu.MenuButton;
-import com.kireiiiiiiii.shooting_stars.ui.elements.menu.MenuScreen;
-import com.kireiiiiiiii.shooting_stars.ui.elements.menu.PopUpPanelWindget;
-import com.kireiiiiiiii.shooting_stars.ui.elements.menu.links_panel.GithubLink;
-import com.kireiiiiiiii.shooting_stars.ui.elements.menu.links_panel.InstagramLink;
-import com.kireiiiiiiii.shooting_stars.ui.elements.menu.settings_panel.ChangeButton;
-import com.kireiiiiiiii.shooting_stars.ui.elements.menu.settings_panel.LanguageTitle;
+import com.kireiiiiiiii.shooting_stars.ui.Backround;
+import com.kireiiiiiiii.shooting_stars.ui.game.GameOverScreen;
+import com.kireiiiiiiii.shooting_stars.ui.game.HomeButton;
+import com.kireiiiiiiii.shooting_stars.ui.game.PauseScreen;
+import com.kireiiiiiiii.shooting_stars.ui.game.ScoreBoard;
+import com.kireiiiiiiii.shooting_stars.ui.game.ScoreWidget;
+import com.kireiiiiiiii.shooting_stars.ui.game.StarWidget;
+import com.kireiiiiiiii.shooting_stars.ui.game.TimerWidget;
+import com.kireiiiiiiii.shooting_stars.ui.game.TopscoreWidget;
+import com.kireiiiiiiii.shooting_stars.ui.menu.MenuButton;
+import com.kireiiiiiiii.shooting_stars.ui.menu.MenuScreen;
+import com.kireiiiiiiii.shooting_stars.ui.menu.PopUpPanelWindget;
+import com.kireiiiiiiii.shooting_stars.ui.menu.links_panel.GithubLink;
+import com.kireiiiiiiii.shooting_stars.ui.menu.links_panel.InstagramLink;
+import com.kireiiiiiiii.shooting_stars.ui.menu.settings_panel.ChangeButton;
+import com.kireiiiiiiii.shooting_stars.ui.menu.settings_panel.LanguageTitle;
 
 /**
  * Main game object.
