--- src/main/java/com/kireiiiiiiii/shooting_stars/ui/GameScreenMode.java
@@ -1,37 +1,0 @@
-/*
- * Author: Matěj Šťastný
- * Date created: 5/13/2024
- * Github link: https://github.com/kireiiiiiiii/ShootingStars
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- */
-
-package com.kireiiiiiiii.shooting_stars.ui;
-
-/**
- * Enum class to handle window type to be painted on the {@code JPanel}.
- * 
- */
-public enum GameScreenMode {
-    GAME,
-    PAUSE,
-    GAME_OVER
-}
