--- src/main/java/com/kireiiiiiiii/shooting_stars/ui/MenuScreenMode.java
@@ -1,35 +1,0 @@
-/*
- * Author: Matěj Šťastný
- * Date created: 6/16/2024
- * Github link: https://github.com/kireiiiiiiii/ShootingStars
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- */
-
-package com.kireiiiiiiii.shooting_stars.ui;
-
-/**
- * Enum class to handle window type to be painted on the {@code JPanel}.
- * 
- */
-public enum MenuScreenMode {
-    MAIN, SETTINGS, LINKS
-}
