+++ src/main/java/com/kireiiiiiiii/shooting_stars/ui/game/StarWidget.java
@@ -24,7 +24,7 @@
  *
  */
 
-package com.kireiiiiiiii.shooting_stars.ui.elements.game;
+package com.kireiiiiiiii.shooting_stars.ui.game;
 
 import java.awt.Graphics2D;
 import java.awt.Container;
@@ -36,14 +36,14 @@ import com.kireiiiiiiii.shooting_stars.constants.Interact;
 import com.kireiiiiiiii.shooting_stars.constants.Textures;
 import com.kireiiiiiiii.shooting_stars.constants.WidgetTags;
 import com.kireiiiiiiii.shooting_stars.constants.ZIndexes;
+import com.kireiiiiiiii.shooting_stars.interfaces.Renderable;
 import com.kireiiiiiiii.shooting_stars.tools.ImageUtil;
-import com.kireiiiiiiii.shooting_stars.ui.Renderable;
 
 /**
  * The target widget.
  * 
  */
-public class StarWidget implements Renderable, com.kireiiiiiiii.shooting_stars.ui.Interactable {
+public class StarWidget implements Renderable, com.kireiiiiiiii.shooting_stars.interfaces.Interactable {
 
     /////////////////
     // Constants
