+++ src/main/java/com/kireiiiiiiii/shooting_stars/ui/menu/settings_panel/LanguageTitle.java
@@ -24,7 +24,7 @@
  *
  */
 
-package com.kireiiiiiiii.shooting_stars.ui.elements.menu.settings_panel;
+package com.kireiiiiiiii.shooting_stars.ui.menu.settings_panel;
 
 import java.awt.Color;
 import java.awt.Font;
@@ -36,8 +36,8 @@ import com.kireiiiiiiii.shooting_stars.constants.Fonts;
 import com.kireiiiiiiii.shooting_stars.constants.GameDialogue;
 import com.kireiiiiiiii.shooting_stars.constants.WidgetTags;
 import com.kireiiiiiiii.shooting_stars.constants.ZIndexes;
+import com.kireiiiiiiii.shooting_stars.interfaces.Renderable;
 import com.kireiiiiiiii.shooting_stars.tools.FontUtil;
-import com.kireiiiiiiii.shooting_stars.ui.Renderable;
 
 /**
  * Widget to display the current game language.
